Book Recommendation/bookRecommend/requirements.txt
astroid==1.5.3
colorama==0.3.9
Django==1.9.1
django-braces==1.8.1
django-cors-headers==1.1.0
django-model-utils==2.5.2
django-oauth-toolkit==0.10.0
django-rest-swagger==0.3.4
djangorestframework==3.3.2
docutils==0.12
futures==3.0.5
jmespath==0.9.0
lazy-object-proxy==1.2.2
numpy==1.13.3
oauthlib==1.0.3
pandas==0.21.0
pylint==1.5.0
python-dateutil==2.6.0
pytz==2017.3
PyYAML==3.12
requests==2.9.1
scipy==1.0.0
six==1.11.0
wrapt==1.10.11
